style(subtitle): rename subtitle helpers
